feat: add proving system relations (#96)
Description
---
Adds proving system relations to
[RFC-0182](https://rfc.tari.com/RFC-0182_CommitmentSignatures.html).

Motivation and Context
---
The recent addition of RFC-0182 includes details for commitment and
public key signatures. These include interactive protocols, descriptions
of the use of the Fiat-Shamir technique for non-interactivity, and
security proofs. However, the RFC did not specifically list the proving
system relations used for the two protocols, instead describing them
more informally in the introduction.

This PR adds the proving system relations for improved clarity.

How Has This Been Tested?
---
The changes build and appear to render correctly.