fix: coin base rules (#99)
Description
---
Current implementation of the coinbase builder makes these 0, but we do
not have a rule that specifies they have to be 0.

Motivation and Context
---
If they are not zero, you cannot do validation on the coinbase kernel. 
The script_offset makes zero difference on the malleability of the
kernel so its not required. There is also no spending script key,
meaning you leak the private key to sender_offset key.
The transactions offset makes it impossible to test the value of the
kernel in any block that has more than the coinbase transaction.