feat!: update stablecoin specification
This is a much more fully fledged design for a privacy-enabled
stablecoin and one that approaches feasibility.

It is still likely insecure in places and there are still a few holes to
plug in some of the operations, but the broad brushstrokes and intent is
certainly there.

The specifcation given here focuses on the surveillance-enabled
stablecoin (i.e. CBDC) model. Since this is actually the more difficult
case to implement in a zero-knowledge manner, it is fairly straightforward to modify the specification to implement a free (as in speech) stablecoin (i.e. digital cash) model.

Some hints are given at the end on how one might actually achieve this.
There are also some tweaks that could allow/deny the central issuer to
arbitrarily seize funds from accounts. Obviously a free model would
disallow this feature.

This is still a draft proposal, and as such, I expect there to be
several more revisions before it becomes a viable design. Hopefully this
is a big step in the right direction.