fix: remove dotenv
dotenv was used for testing of the `build.js` script and is not needed
anymore.
Removing it for security reasons. There weren't any major concerns but
some fat-finger errors may have led to deployment problems.
The envars are set in the build environments (e.g. cloudflare) or have
sane fallback defaults.